requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
dynaconf==3.1.12
executing==1.2.0
faiss-cpu==1.7.3
frozenlist==1.3.3
gptcache==0.1.10
greenlet==2.0.2
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.139
llminterface @ git+https://github.com/garland3/llminterface.git@34435680b3eb5255e6b15be584090d824130b2f4
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.0
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.3.3
toml==0.10.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
